feat: add playbackRangeStart and playbackRangeEnd props
Props allow setting a playback range to only load a play a portion of the video file.

BREAKING CHANGE: Fully removes hoverTargetRef prop

re #45
